+++ src/components/book-image/__tests__/book-image-test-cases.js
@@ -1,0 +1,13 @@
+import BookImage from '../book-image';
+
+const testCases = {};
+
+testCases.basic = {
+  component: BookImage,
+  description: 'Basic',
+  props: {
+    size: '60'
+  }
+};
+
+export { testCases };